Normalize theme file paths and rename tests for consistency; add unit tests.
