+++ tests/unit/api/transformer/TransformerTest.php
@@ -9,7 +9,7 @@ use LimeSurvey\Api\Transformer\Transformer;
 /**
  * @testdox API Transformer
  */
-class TransformerOutputTest extends TestBaseClass
+class TransformerTest extends TestBaseClass
 {
     /**
      * @testdox transform() Transformed data only includes fields specified in map.
