+++ tests/unit/helpers/ExpressionCoreFunctionsTest.php
@@ -8,7 +8,7 @@ namespace ls\tests;
  * @since 2022-09-30
  * @group em
  */
-class ExpressionCoreFunctionTest extends TestBaseClass
+class ExpressionCoreFunctionsTest extends TestBaseClass
 {
     /**
      * @group me
